GTFS: Transcoding to GraphTiles (#3669)
* first commit

* comments

* submodule sync

* serialized nodes

* starting building tiles

* starting building tiles

* serialized shapes

* serialized shapes

* move functions to read_transit

* move functions to read_transit

* debugging ingest_transit

* debugging complete - ready to test!

* formatting and comments

* formatting bug fixed

* fix missing includes

* .pbf files are properly written

* able to loop through multiple feeds

* remove geos

* minor fixes

* convert transit include errors

* debug level = info

* remove geos

* threading test error fix

* suggested minor fixes

* suggested minor fixes

* minor changes

* update changelog

* syntax

* syntax error

* fixes

* testing for stitching
