+++ scripts/valhalla_build_config
@@ -31,6 +31,7 @@ config = {
     'admin': '/data/valhalla/admin.sqlite',
     'timezone': '/data/valhalla/tz_world.sqlite',
     'transit_dir': '/data/valhalla/transit',
+    'transit_feeds_dir': '/data/valhalla/transit_feeds',
     'transit_bounding_box': Optional(str),
     'hierarchy': True,
     'shortcuts': True,
@@ -301,6 +302,7 @@ help_text = {
     'admin': 'Location of sqlite file holding admin polygons created with valhalla_build_admins',
     'timezone': 'Location of sqlite file holding timezone information created with valhalla_build_timezones',
     'transit_dir': 'Location of intermediate transit tiles created with valhalla_build_transit',
+    'transit_feeds_dir': 'Location of GTFS transit feeds',
     'transit_bounding_box': 'Add comma separated bounding box values to only download transit data inside the given bounding box',
     'hierarchy': 'bool indicating whether road hierarchy is to be built - default to True',
     'shortcuts': 'bool indicating whether shortcuts are to be built - default to True',
